agent_service/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
requests
pymongo
httpx
# AI & RAG
torch
Crawl4AI
transformers
datasets
huggingface_hub
sentence-transformers
langgraph
langchain
langchain-community
langchain-groq
langgraph-checkpoint-mongodb
langchain-google-genai
langchain-ollama
ollama
langchain-tavily
beautifulsoup4
duckduckgo-search
ddgs
tavily-python
faiss-cpu
einops
# Utilities
numpy
pandas
tqdm
boto3